Allow IO use in the solver.
This simply adds IO to the return type of the `DependencyResolver`
type synonym. There's still a `Progress` type being returned. One
can argue that if solvers are given the opportunity to do IO
themselves, they should also take care of printing their own
debugging information, and `Progress` should disappear from the
output. We'll leave such a change for the future.

The motivation of the IO change is to allow solvers that interface
with external processes or impure libraries such as SMT solvers.